# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/courses/{slug}": {
      "get": {
        "operationId": "getCourseBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Course not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a published course by slug",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/v1/admin/courses": {
      "get": {
        "operationId": "adminListCourses",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseSummaryDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "workos": []
          }
        ],
        "summary": "List all courses",
        "tags": [
          "admin: courses"
        ]
      },
      "post": {
        "operationId": "adminCreateCourse",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; `errors` maps field paths to messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Slug already in use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "workos": []
          }
        ],
        "summary": "Create a course",
        "tags": [
          "admin: courses"
        ]
      }
    },
    "/api/v1/admin/courses/{id}": {
      "get": {
        "operationId": "adminGetCourse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDetailDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Course not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "workos": []
          }
        ],
        "summary": "Get a course with everything attached",
        "tags": [
          "admin: courses"
        ]
      },
      "patch": {
        "operationId": "adminUpdateCourse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; `errors` maps field paths to messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Course not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Slug already in use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "workos": []
          }
        ],
        "summary": "Update a course",
        "tags": [
          "admin: courses"
        ]
      },
      "delete": {
        "operationId": "adminDeleteCourse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Course not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "workos": []
          }
        ],
        "summary": "Delete a course",
        "tags": [
          "admin: courses"
        ]
      }
    },
    "/api/v1/admin/courses/{id}/curriculum": {
      "put": {
        "operationId": "adminReplaceCurriculum",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceCurriculumDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CurriculumModuleDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; `errors` maps field paths to messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Course not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "workos": []
          }
        ],
        "summary": "Replace the whole curriculum module list",
        "tags": [
          "admin: courses"
        ]
      }
    },
    "/api/v1/admin/courses/{id}/teachers": {
      "put": {
        "operationId": "adminReplaceCourseTeachers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceCourseTeachersDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeacherDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; `errors` maps field paths to messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Course not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "workos": []
          }
        ],
        "summary": "Replace the teachers assigned to a course",
        "tags": [
          "admin: courses"
        ]
      }
    },
...
  },
  "info": {
    "title": "PrincetonBD API",
    "description": "Backend API for the PrincetonBD web application.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
  "components": {
    "securitySchemes": {
      "workos": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "WorkOS AuthKit access token."
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "Course not found"
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "ValidationErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "Course not found"
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {
              "name": [
                "name should not be empty"
              ],
              "modules.0.title": [
                "title must be a string"
              ]
            }
          }
        },
        "required": [
          "statusCode",
          "message",
          "errors"
        ]
      },
...
      "CourseSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "example": "sat"
          },
          "name": {
            "type": "string",
            "example": "SAT"
          },
          "description": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "priceAmount": {
            "type": "number",
            "description": "Whole taka.",
            "example": 45000
          },
          "priceUnit": {
            "type": "string",
            "example": "per 10-week course"
          },
          "currency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "modes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeliveryMode"
            }
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CourseStatus"
              }
            ]
          },
          "durationWeeks": {
            "type": "number",
            "nullable": true
          },
          "taughtHours": {
            "type": "number",
            "nullable": true
          },
          "mockCount": {
            "type": "number",
            "nullable": true
          },
          "classSize": {
            "type": "string",
            "nullable": true,
            "example": "Max 10"
          },
          "sortOrder": {
            "type": "number"
          },
          "nextBatch": {
            "nullable": true,
            "description": "The soonest batch a learner can still join, or null when none is scheduled.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/BatchDto"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "thumbnailUrl",
          "priceAmount",
          "priceUnit",
          "currency",
          "modes",
          "status",
          "durationWeeks",
          "taughtHours",
          "mockCount",
          "classSize",
          "sortOrder",
          "nextBatch",
          "createdAt",
          "updatedAt"
        ]
      },
      "CurriculumModuleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "position": {
            "type": "number",
            "description": "1-based display order.",
            "example": 1
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "topics": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "hours": {
            "type": "number",
            "nullable": true
          },
          "outcome": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "position",
          "title",
          "summary",
          "topics",
          "hours",
          "outcome"
        ]
      },
      "TeacherDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "example": "farzana-haque"
          },
          "name": {
            "type": "string",
            "example": "Farzana Haque"
          },
          "designation": {
            "type": "string",
            "description": "The title shown under the name.",
            "example": "Lead Instructor, Quantitative"
          },
          "bio": {
            "type": "string",
            "description": "One or two lines: the credential that matters."
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "description": "Public URL of an uploaded photo; null renders a monogram."
          },
          "branch": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/BranchDto"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "designation",
          "bio",
          "imageUrl",
          "branch",
          "isActive",
          "sortOrder",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CourseDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "example": "sat"
          },
          "name": {
            "type": "string",
            "example": "SAT"
          },
          "description": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "priceAmount": {
            "type": "number",
            "description": "Whole taka.",
            "example": 45000
          },
          "priceUnit": {
            "type": "string",
            "example": "per 10-week course"
          },
          "currency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "modes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeliveryMode"
            }
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CourseStatus"
              }
            ]
          },
          "durationWeeks": {
            "type": "number",
            "nullable": true
          },
          "taughtHours": {
            "type": "number",
            "nullable": true
          },
          "mockCount": {
            "type": "number",
            "nullable": true
          },
          "classSize": {
            "type": "string",
            "nullable": true,
            "example": "Max 10"
          },
          "sortOrder": {
            "type": "number"
          },
          "nextBatch": {
            "nullable": true,
            "description": "The soonest batch a learner can still join, or null when none is scheduled.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/BatchDto"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "feeIncludes": {
            "description": "Bullet list on the fee card.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "outcomes": {
            "description": "\"By the end of the course you will…\"",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CurriculumModuleDto"
            }
          },
          "batches": {
            "description": "Publicly: upcoming, joinable batches soonest first. For admins: every batch.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BatchDto"
            }
          },
          "teachers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeacherDto"
            }
          },
          "testimonials": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestimonialDto"
            }
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "thumbnailUrl",
          "priceAmount",
          "priceUnit",
          "currency",
          "modes",
          "status",
          "durationWeeks",
          "taughtHours",
          "mockCount",
          "classSize",
          "sortOrder",
          "nextBatch",
          "createdAt",
          "updatedAt",
          "feeIncludes",
          "outcomes",
          "modules",
          "batches",
          "teachers",
          "testimonials"
        ]
      },
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "description": "URL segment under /test-prep/. Derived from `name` when omitted.",
            "maxLength": 80,
            "example": "sat",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "example": "SAT"
          },
          "description": {
            "type": "string",
            "description": "One or two sentences; shown on cards and used as the default page intro.",
            "maxLength": 2000
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true,
            "description": "Public URL from the upload endpoint. Null clears it.",
            "maxLength": 1024,
            "format": "uri"
          },
          "priceAmount": {
            "type": "number",
            "description": "Whole taka.",
            "minimum": 0,
            "example": 45000
          },
          "priceUnit": {
            "type": "string",
            "maxLength": 80,
            "example": "per 10-week course"
          },
          "feeIncludes": {
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 200
            }
          },
          "modes": {
            "type": "array",
            "uniqueItems": true,
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/DeliveryMode"
            }
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CourseStatus"
              }
            ]
          },
          "durationWeeks": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "taughtHours": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "mockCount": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "classSize": {
            "type": "string",
            "nullable": true,
            "maxLength": 40,
            "example": "Max 10"
          },
          "outcomes": {
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 300
            }
          },
          "sortOrder": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "name",
          "priceAmount",
          "modes"
        ]
      },
      "UpdateCourseDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "description": "URL segment under /test-prep/. Derived from `name` when omitted.",
            "maxLength": 80,
            "example": "sat",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "example": "SAT"
          },
          "description": {
            "type": "string",
            "description": "One or two sentences; shown on cards and used as the default page intro.",
            "maxLength": 2000
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true,
            "description": "Public URL from the upload endpoint. Null clears it.",
            "maxLength": 1024,
            "format": "uri"
          },
          "priceAmount": {
            "type": "number",
            "description": "Whole taka.",
            "minimum": 0,
            "example": 45000
          },
          "priceUnit": {
            "type": "string",
            "maxLength": 80,
            "example": "per 10-week course"
          },
          "feeIncludes": {
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 200
            }
          },
          "modes": {
            "type": "array",
            "uniqueItems": true,
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/DeliveryMode"
            }
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CourseStatus"
              }
            ]
          },
          "durationWeeks": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "taughtHours": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "mockCount": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "classSize": {
            "type": "string",
            "nullable": true,
            "maxLength": 40,
            "example": "Max 10"
          },
          "outcomes": {
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 300
            }
          },
          "sortOrder": {
            "type": "number",
            "example": 0
          }
        }
      },
...
      "ReplaceCurriculumDto": {
        "type": "object",
        "properties": {
          "modules": {
            "maxItems": 40,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CurriculumModuleInputDto"
            }
          }
        },
        "required": [
          "modules"
        ]
      },
      "ReplaceCourseTeachersDto": {
        "type": "object",
        "properties": {
          "teacherIds": {
            "uniqueItems": true,
            "maxItems": 50,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "teacherIds"
        ]
      },
...
    }
  }
}